refactor: move JIM.Application interfaces to Interfaces folder (#394)
Move ISyncServer, IJimApplicationFactory, and ICredentialProtectionService
to src/JIM.Application/Interfaces/ with namespace JIM.Application.Interfaces,
matching the convention used in JIM.Models/Interfaces/.

Add using JIM.Application.Interfaces to all consumers across source and
test projects. Add to _Imports.razor for Blazor pages.
