+++ src/JIM.Application/JimApplication.cs
@@ -1,3 +1,4 @@
+using JIM.Application.Interfaces;
 using JIM.Application.Servers;
 using JIM.Application.Services;
 using JIM.Data;
