+++ src/JIM.Web/Program.cs
@@ -2,6 +2,7 @@ using System.Text.Json;
 using Asp.Versioning;
 using JIM.Application;
 using JIM.Application.Expressions;
+using JIM.Application.Interfaces;
 using JIM.Application.Services;
 using JIM.Data;
 using JIM.Web.Models;
