+++ src/JIM.Web/_Imports.razor
@@ -1,4 +1,5 @@
 ﻿@using JIM.Application
+@using JIM.Application.Interfaces
 @using JIM.Web
 @using JIM.Web.Services
 @using JIM.Web.Shared
