+++ src/JIM.Worker/Processors/SyncExportTaskProcessor.cs
@@ -1,5 +1,6 @@
 using JIM.Application;
 using JIM.Application.Diagnostics;
+using JIM.Application.Interfaces;
 using JIM.Application.Utilities;
 using JIM.Connectors;
 using JIM.Connectors.File;
