+++ src/JIM.Worker/Worker.cs
@@ -1,5 +1,6 @@
 using JIM.Application;
 using JIM.Application.Diagnostics;
+using JIM.Application.Interfaces;
 using JIM.Connectors;
 using JIM.Models.Activities;
 using JIM.Models.Core;
