+++ test/JIM.Web.Api.Tests/SynchronisationControllerClearTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using JIM.Application;
+using JIM.Application.Interfaces;
 using JIM.Application.Expressions;
 using JIM.Application.Services;
 using JIM.Data;
