+++ test/JIM.Web.Api.Tests/SynchronisationControllerRunProfileTests.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using JIM.Web.Controllers.Api;
 using JIM.Web.Models.Api;
 using JIM.Application;
+using JIM.Application.Interfaces;
 using JIM.Application.Expressions;
 using JIM.Models.Interfaces;
 using JIM.Application.Services;
